// cpo-server/src/types/Transaction.ts
export enum InactivityStatus {
  INFO = 'I',
  WARNING = 'W',
  ERROR = 'E'
}
// ...
export default interface Transaction extends AbstractCurrentConsumption, TransactionAuthorizationActions {
  id?: number;
  carID?: string;
  car?: Car;
  carCatalogID?: number;
  carCatalog?: CarCatalog;
  carStateOfCharge?: number;
  carOdometer?: number;
  departureTime?: Date;
  targetStateOfCharge?: number;
  phasesUsed?: CSPhasesUsed;
  companyID?: string;
  company?: Company;
  siteID?: string;
  site?: Site;
  siteAreaID?: string;
  siteArea?: SiteArea;
  issuer: boolean;
  connectorId: number;
  tagID: string;
  emaidID?: string;
  tag?: Tag;
  emaid?:Emaid;
  userID: string;
  chargeBoxID: string;
  signedData?: string;
  user?: User;
  stop?: TransactionStop;
  remotestop?: {
    timestamp: Date;
    tagID: string;
    userID: string;
  };
  chargeBox?: ChargingStation;
  meterStart: number;
  timestamp: Date;
  price?: number;
  roundedPrice?: number;
  priceUnit?: string;
  pricingSource?: string;
  pricingModel?: ResolvedPricingModel,
  stateOfCharge: number;
  timezone: string;
  currentTimestamp?: Date;
  currentTotalInactivitySecs: number;
  currentInactivityStatus?: InactivityStatus;
  currentStateOfCharge: number;
  currentTotalDurationSecs?: number;
  transactionEndReceived?: boolean;
  currentCumulatedPrice?: number;
  currentCumulatedRoundedPrice?: number; // added to address rounding issues on multiple pricing dimension
  currentSignedData?: string;
  numberOfMeterValues: number;
  uniqueId?: string;
  values?: Consumption[];
  ocpi?: boolean;
  ocpiWithCdr?: boolean;
  ocpiData?: TransactionOcpiData;
  oicpData?: TransactionOicpData;
  billingData?: TransactionBillingData;
  refundData?: TransactionRefundData;
  migrationTag?: string;
  authorizationID?: string;
  tariffData?:ResolvedPricingDefinition[];
}

export interface TransactionStats extends DatabaseCount {
  totalConsumptionWattHours?: number;
  totalPriceRefund?: number;
  totalPricePending?: number;
  countRefundTransactions?: number;
  countPendingTransactions?: number;
  countRefundedReports?: number;
  totalDurationSecs?: number;
  totalPrice?: number;
  currency?: string;
  totalInactivitySecs?: number;
}

export interface TransactionOcpiData {
  session?: OCPISession;
  cdr?: OCPICdr;
  sessionCheckedOn?: Date;
  cdrCheckedOn?: Date;
}

export interface TransactionOicpData {
  session?: OICPSession;
  cdr?: OICPChargeDetailRecord;
  sessionCheckedOn?: Date;
  cdrCheckedOn?: Date;
}

export interface CSPhasesUsed {
  csPhase1: boolean;
  csPhase2: boolean;
  csPhase3: boolean;
}

export interface TransactionStop {
  timestamp: Date;
  meterStop: number;
  reason?: string;
  tagID: string;
  userID: string;
  user?: User;
  price?: number;
  roundedPrice?: number;
  priceUnit?: string;
  pricingSource?: string;
  stateOfCharge?: number;
  totalInactivitySecs?: number;
  extraInactivitySecs?: number;
  extraInactivityComputed?: boolean;
  totalConsumptionWh?: number;
  totalDurationSecs?: number;
  inactivityStatus?: InactivityStatus;
  transactionData?: OCPP15TransactionData|OCPPMeterValue[];
  signedData?: string;
}